Refactored tests to run in sub directories
