+++ application/tests/__init__.py
@@ -1,0 +1,5 @@
+from .tests_application import *
+from .tests_business_logic import *
+from .tests_validation import *
+from .tests_views import *
+from .tests_selenium import *
