--- application/tests/tests_validation/__init__.py
@@ -1,4 +1,0 @@
-from .test_user_details import *
-from .test_dbs_check import *
-from .test_payment import *
-from .test_personal_details import *
