+++ application/tests/tests_views/test_application_saved.py
@@ -1,4 +1,4 @@
-from .test_views import *
+from .view_parent import *
 
 
 class ApplicationSavedTest(ViewsTest):
