+++ application/tests/tests_views/test_dbs_check.py
@@ -1,4 +1,4 @@
-from .test_views import *
+from .view_parent import *
 
 
 class DBSCheckTest(ViewsTest):
