+++ application/tests/tests_views/test_first_aid_training.py
@@ -1,4 +1,4 @@
-from .test_views import *
+from .view_parent import *
 
 
 class FirstAidTrainingTest(TestCase):
