+++ application/tests/tests_views/test_health.py
@@ -1,4 +1,4 @@
-from .test_views import *
+from .view_parent import *
 
 
 class HealthTest(ViewsTest):
