+++ application/tests/tests_views/test_other_people.py
@@ -1,4 +1,4 @@
-from .test_views import *
+from .view_parent import *
 
 
 class OtherPeopleTest(ViewsTest):
