+++ application/tests/tests_views/test_task_status.py
@@ -1,4 +1,4 @@
-from .test_views import *
+from .view_parent import *
 
 
 class TaskStatusTest(ViewsTest):
