+++ application/tests/tests_views/test_type_of_childcare.py
@@ -1,4 +1,4 @@
-from .test_views import *
+from .view_parent import *
 
 
 class TypeOfChildcareTest(ViewsTest):
