+++ application/tests/tests_views/test_validate_magic_link_email.py
@@ -1,4 +1,4 @@
-from .test_views import *
+from .view_parent import *
 
 
 class ValidateMagicLinkEmailTest(ViewsTest):
